requirements.txt
fastapi
uvicorn[standard]
starlette
sqlalchemy
alembic
redis
pydantic[email]
python-dotenv
psycopg2-binary
typing-extensions
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiohttp
jinja2
feedparser
fastapi-mail
feedgen